backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.7.1
python-multipart==0.0.9
httpx==0.27.2
lxml
numpy
beautifulsoup4
pydantic-settings
openai>=1.0.0
google-cloud-vision>=3.0.0
google-cloud-aiplatform>=1.35.0
